fix(analyzers/general): IMPROVE MI general analyzer
